Completed UI for register page as well.
