+++ QRCodeReader/LoginScreen.swift
@@ -20,30 +20,12 @@ class LoginScreen: UIViewController {
     @IBOutlet weak var email: UITextField!
     
     @IBOutlet weak var password: UITextField!
-    struct Colors {
-        
-        static let brightOrange     = UIColor(red: 255.0/255.0, green: 69.0/255.0, blue: 0.0/255.0, alpha: 1.0)
-        static let red              = UIColor(red: 255.0/255.0, green: 115.0/255.0, blue: 115.0/255.0, alpha: 1.0)
-        static let orange           = UIColor(red: 255.0/255.0, green: 175.0/255.0, blue: 72.0/255.0, alpha: 1.0)
-        static let blue             = UIColor(red: 74.0/255.0, green: 144.0/255.0, blue: 228.0/255.0, alpha: 1.0)
-        static let green            = UIColor(red: 91.0/255.0, green: 197.0/255.0, blue: 159.0/255.0, alpha: 1.0)
-        static let darkGrey         = UIColor(red: 85.0/255.0, green: 85.0/255.0, blue: 85.0/255.0, alpha: 1.0)
-        static let veryDarkGrey     = UIColor(red: 13.0/255.0, green: 13.0/255.0, blue: 13.0/255.0, alpha: 1.0)
-        static let lightGrey        = UIColor(red: 200.0/255.0, green: 200.0/255.0, blue: 200.0/255.0, alpha: 1.0)
-        static let black            = UIColor(red: 0.0/255.0, green: 0.0/255.0, blue: 0.0/255.0, alpha: 1.0)
-        static let white            = UIColor(red: 255.0/255.0, green: 255.0/255.0, blue: 255.0/255.0, alpha: 1.0)
-    }
+
     
     @IBAction func exitTextField(_ sender: Any) {
         (sender as! UITextField).resignFirstResponder()
     }
     
-    func setHeight(textField: UITextField, height: Int) {
-        var rect = textField.frame
-        rect.size.height = 45; // <-- Specify the height you want here.
-        
-        textField.frame = rect
-    }
     
     
     override func viewDidLoad() {
