+++ QRCodeReader/RegisterScreen.swift
@@ -1,0 +1,41 @@
+//
+//  RegisterScreen.swift
+//  QRCodeReader
+//
+//  Created by Samuel Lefcourt on 1/31/19.
+//  Copyright © 2019 AppCoda. All rights reserved.
+//
+
+import Foundation
+import UIKit
+import AVFoundation
+class RegisterScreen: UIViewController {
+    @IBOutlet weak var cPassword: UITextField!
+    @IBOutlet weak var password: UITextField!
+    @IBOutlet weak var email: UITextField!
+    @IBOutlet weak var registerButton: UIButton!
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        self.view.addGestureRecognizer(UITapGestureRecognizer(target: self.view, action: #selector(UIView.endEditing(_:))))
+        view.setGradientBackground(colorOne: Colors.lightGrey, colorTwo: Colors.green)
+        registerButton.layer.cornerRadius = registerButton.frame.size.height/2
+        registerButton.layer.masksToBounds = true
+        
+        if let myImage = UIImage(named: "email"){
+            email.withImage(direction: .Left, image: myImage, colorSeparator: UIColor.orange, colorBorder: UIColor.black)
+            setHeight(textField: email, height: 45)
+        }
+        
+        if let myImage = UIImage(named: "lock"){
+            password.withImage(direction: .Left, image: myImage, colorSeparator: UIColor.orange, colorBorder: UIColor.black)
+            setHeight(textField: password, height: 45)
+        }
+        
+        if let myImage = UIImage(named: "lock"){
+            cPassword.withImage(direction: .Left, image: myImage, colorSeparator: UIColor.orange, colorBorder: UIColor.black)
+            setHeight(textField: cPassword, height: 45)
+        }
+        
+    }
+    
+}
